refactor: Presenterへ処理を集約
