+++ Runtime/VoiceMimicModel.cs
@@ -1,9 +1,7 @@
 namespace VoiceMimic
 {
     using System;
-    using System.Text;
     using System.Collections.Generic;
-    using System.IO;
     using System.Linq;
     using UnityEngine;
 
@@ -225,47 +223,6 @@ namespace VoiceMimic
             }
 
             return new PcmBuffer { sampleRate = snap.sampleRate, channels = channels, samples = interleaved };
-        }
-
-        public void ExportWav(PcmBuffer pcm, string path)
-        {
-            if (string.IsNullOrEmpty(path))
-            {
-                throw new ArgumentException("出力パスが指定されていません", path);
-            }
-
-            var dir = Path.GetDirectoryName(path);
-            if (string.IsNullOrEmpty(dir) == false && Directory.Exists(dir) == false)
-            {
-                Directory.CreateDirectory(dir);
-            }
-
-            using var fs = new FileStream(path, FileMode.Create, FileAccess.Write);
-            using var bw = new BinaryWriter(fs);
-            var dataLength = pcm.samples.Length * 2;
-            var riffLength = 36 + dataLength;
-
-            bw.Write(Encoding.ASCII.GetBytes("RIFF"));
-            bw.Write(riffLength);
-            bw.Write(Encoding.ASCII.GetBytes("WAVE"));
-
-            bw.Write(Encoding.ASCII.GetBytes("fmt "));
-            bw.Write(16);
-            bw.Write((short)1);
-            bw.Write((short)pcm.channels);
-            bw.Write(pcm.sampleRate);
-            bw.Write(pcm.sampleRate * pcm.channels * 2);
-            bw.Write((short)(pcm.channels * 2));
-            bw.Write((short)16);
-
-            bw.Write(Encoding.ASCII.GetBytes("data"));
-            bw.Write(dataLength);
-
-            for (int i = 0; i < pcm.samples.Length; i++)
-            {
-                var s = (short)Mathf.Clamp(pcm.samples[i] * 32767f, -32768f, 32767f);
-                bw.Write(s);
-            }
         }
 
         public SequenceSnapshot SnapshotFromAsset(VoiceMimicAsset asset)
